.env implementado configuraciones en config.js TODO: logout

+++ src/db/migrations/20200202230133-create-user.js
@@ -6,7 +6,7 @@ module.exports = {
         allowNull: false,
         autoIncrement: true,
         primaryKey: true,
-        type: Sequelize.INTEGER(11)
+        type: Sequelize.INTEGER
       },
       name: {
         type: Sequelize.STRING
